+++ src/components/emoji-picker.tsx
@@ -509,6 +509,23 @@ const EmojiPickerSearch = forwardRef<HTMLInputElement, EmojiPickerSearchProps>(
         store.set({ searchRef: ref });
       }
     }, []);
+    const isControlled = typeof value === "string";
+    const wasControlled = useRef(isControlled);
+
+    useEffect(() => {
+      if (
+        process.env.NODE_ENV !== "production" &&
+        wasControlled.current !== isControlled
+      ) {
+        console.warn(
+          `EmojiPicker.Search is changing from ${
+            wasControlled ? "controlled" : "uncontrolled"
+          } to ${isControlled ? "controlled" : "uncontrolled"}.`,
+        );
+      }
+
+      wasControlled.current = isControlled;
+    }, [isControlled]);
 
     // Initialize search with a controlled or uncontrolled value
     useLayoutEffect(() => {
@@ -521,6 +538,13 @@ const EmojiPickerSearch = forwardRef<HTMLInputElement, EmojiPickerSearchProps>(
               : "",
       });
     }, []);
+
+    // Handle controlled value changes
+    useLayoutEffect(() => {
+      if (typeof value === "string") {
+        store.get().onSearchChange(value);
+      }
+    }, [value]);
 
     const handleChange = useCallback(
       (event: ReactChangeEvent<HTMLInputElement>) => {